requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
pytesseract==0.3.13
pytest==8.3.3
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
sympy==1.13.3
tokenizers==0.12.1
torch==2.2.2
tqdm==4.66.6
transformers==4.21.0
typing_extensions==4.12.2
urllib3==2.2.3
